+++ mlfinlab/optimal_mean_reversion/__init__.py
@@ -1,5 +1,0 @@
-"""
-Classes derived from Optimal Mean Reversion module.
-"""
-
-from mlfinlab.optimal_mean_reversion.ou_model import OrnsteinUhlenbeck